Windows platform fixes and Cloud Locate/test fix.
MSVC had a few more things that needed coping with:
- MSVC asserts if the argument passed to isprint() is out of the range of a char, so some additional casting is required.
- my slightly dodgy construction to printf() a location as a clickable link in the example code, which required the parameters to be called in order, doesn't work under MSVC and so needed some tidying up,
- ${UBXLIB_PRIVATE_INC} needed to be included when building u_main.c when U_CFG_MUTEX_DEBUG is switched on.

On the Cloud Locate front, the seconds were being misinterpreted when parsing the response string and the unsubscribe from the Cloud Locate topic was not occurring.  On the test front, the location test wasn't selecting configuration data properly when there was more than one set of configuration data per network type and it also wasn't checking for a NULL pointer.

+++ cell/src/u_cell_mqtt.c
@@ -1150,7 +1150,7 @@ static bool isPrint(const char *pBuffer, size_t bufferLength)
     bool printable = true;
 
     for (size_t x = 0; (x < bufferLength) && printable; x++) {
-        if (!isprint((int32_t) *pBuffer)) {
+        if (!isprint((int32_t) (uint8_t) *pBuffer)) {
             printable = false;
         }
         pBuffer++;
